fix: remove 'use client' header
