+++ src/index.ts
@@ -1,3 +1,1 @@
-'use client'
-
 export * from './web'
